+++ CHANGELOG.md
@@ -48,6 +48,8 @@
 - **algorithm/backTrack:** add combine ([74ce4ed](https://github.com/air-supply94/js-algorithm/commit/74ce4eda583e9368d48e1e5948f371ada6064bc8))
 - **algorithm/backTrack:** add solveSudoku ([a85601c](https://github.com/air-supply94/js-algorithm/commit/a85601cb2ed535d46e4efe08673a5188250a5f5e))
 - **algorithm/bfs:** add minDepth ([31ead4a](https://github.com/air-supply94/js-algorithm/commit/31ead4a45634cb14bc5ef9151856e93ee790f421))
+- **algorithm/bfs:** add openLock ([59d2a95](https://github.com/air-supply94/js-algorithm/commit/59d2a95acdbb5b2fb1b2f5eb8fceb9faa2582c64))
+- **algorithm/bfs:** add slidingPuzzle ([6fee149](https://github.com/air-supply94/js-algorithm/commit/6fee1496dec03162235fe58c023a90ddd72089d9))
 - **algorithm/classic:** add coinChange ([4c2b923](https://github.com/air-supply94/js-algorithm/commit/4c2b92300783066c27c01ce3f79989fbdfcf0314))
 - **algorithm/classic:** add maximalSquare ([82da714](https://github.com/air-supply94/js-algorithm/commit/82da714f99bc9b61b8b2498c1bac7e0900b08540))
 - **algorithm/classic:** add maxSubArray ([684c1b8](https://github.com/air-supply94/js-algorithm/commit/684c1b823a8541d96bbbb4e245ef456e09425e35))